backend/requirements.txt
boto3==1.24.61
botocore==1.27.61
click==8.1.3
cycler==0.11.0
Flask==2.2.2
fonttools==4.37.1
gunicorn==20.1.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
numpy==1.23.2
packaging==21.3
pandas==1.2.0
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
s3transfer==0.6.0
scikit-learn==1.1.2
scipy==1.9.1
scorecardpy==0.1.9.2
six==1.16.0
threadpoolctl==3.1.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1